Psalm set errorLevel 1 (max)

+++ src/ResourceStorage.php
@@ -9,6 +9,7 @@ use BEAR\Resource\AbstractUri;
 use BEAR\Resource\RequestInterface;
 use BEAR\Resource\ResourceObject;
 use Doctrine\Common\Cache\CacheProvider;
+use function is_string;
 
 final class ResourceStorage implements ResourceStorageInterface
 {
@@ -45,10 +46,13 @@ final class ResourceStorage implements ResourceStorageInterface
 
     /**
      * {@inheritdoc}
+     *
+     * @return void
      */
     public function updateEtag(ResourceObject $ro, int $lifeTime)
     {
         $varyUri = $this->getVaryUri($ro->uri);
+        assert(isset($ro->headers['ETag']) && is_string($ro->headers['ETag']));
         $etag = self::ETAG_VAL . $ro->headers['ETag'];
         $uri = self::ETAG_TABLE . $varyUri;
         // delete old ETag
@@ -61,13 +65,17 @@ final class ResourceStorage implements ResourceStorageInterface
 
     /**
      * {@inheritdoc}
+     *
+     * @return void
      */
     public function deleteEtag(AbstractUri $uri)
     {
         $varyUri = self::ETAG_TABLE . $this->getVaryUri($uri); // invalidate etag
+        /** @psalm-suppress MixedAssignment */
         $oldEtagKey = $this->cache->fetch($varyUri);
-
-        $this->cache->delete($oldEtagKey);
+        if (is_string($oldEtagKey)) {
+            $this->cache->delete($oldEtagKey);
+        }
     }
 
     /**
@@ -76,8 +84,10 @@ final class ResourceStorage implements ResourceStorageInterface
     public function get(AbstractUri $uri)
     {
         $varyUri = $this->getVaryUri($uri);
+        /** @var array{0:string, 1: int, 2:array<string, string>, 3: mixed, 4: string}|false $roProps */
+        $roProps = $this->cache->fetch($varyUri);
 
-        return $this->cache->fetch($varyUri);
+        return $roProps;
     }
 
     /**
@@ -93,9 +103,12 @@ final class ResourceStorage implements ResourceStorageInterface
 
     /**
      * {@inheritdoc}
+     *
+     * @return bool
      */
     public function saveValue(ResourceObject $ro, int $lifeTime)
     {
+        /** @psalm-suppress MixedAssignment $body */
         $body = $this->evaluateBody($ro->body);
         $uri = $this->getVaryUri($ro->uri);
         $val = [$ro->uri, $ro->code, $ro->headers, $body, null];
@@ -105,9 +118,12 @@ final class ResourceStorage implements ResourceStorageInterface
 
     /**
      * {@inheritdoc}
+     *
+     * @return bool
      */
     public function saveView(ResourceObject $ro, int $lifeTime)
     {
+        /** @psalm-suppress MixedAssignment $body */
         $body = $this->evaluateBody($ro->body);
         $uri = $this->getVaryUri($ro->uri);
         $val = [$ro->uri, $ro->code, $ro->headers, $body, $ro->view];
@@ -115,14 +131,23 @@ final class ResourceStorage implements ResourceStorageInterface
         return $this->cache->save($uri, $val, $lifeTime);
     }
 
+    /**
+     * @param mixed $body
+     *
+     * @return mixed
+     */
     private function evaluateBody($body)
     {
         if (! \is_array($body)) {
             return $body;
         }
+        /** @psalm-suppress MixedAssignment $item */
         foreach ($body as &$item) {
             if ($item instanceof RequestInterface) {
                 $item = ($item)();
+            }
+            if ($item instanceof ResourceObject) {
+                $item->body = $this->evaluateBody($item->body);
             }
         }
 
@@ -134,15 +159,17 @@ final class ResourceStorage implements ResourceStorageInterface
         if (! isset($_SERVER['X_VARY'])) {
             return (string) $uri;
         }
-        $varys = \explode(',', $_SERVER['X_VARY']);
+        /** @var string $xvary */
+        $xvary = $_SERVER['X_VARY'];
+        $varys = \explode(',', $xvary);
         $varyId = '';
         foreach ($varys as $vary) {
             $phpVaryKey = \sprintf('X_%s', \strtoupper($vary));
-            if (isset($_SERVER[$phpVaryKey])) {
+            if (isset($_SERVER[$phpVaryKey]) && is_string($_SERVER[$phpVaryKey])) {
                 $varyId .= $_SERVER[$phpVaryKey];
             }
         }
 
-        return (string) $uri . $varyId;
+        return $uri . $varyId;
     }
 }
